Do not upgrade setuptools to the latest version before installing gophish-tools
We don't upgrade setuptools here because the versions of gophish on
PYPI are ancient and depend on an equally ancient version of
setuptools.  The gophish package installation will fail if we let pip
pull in the latest setuptools; instead we let the gophish package
installation itself choose the version of setuptools to install.